django_project/requirements.txt
django
crispy-bootstrap4
django-crispy-forms
djangorestframework